Refactoring member access
